test: 예약 생성 서비스 클래스 분리
