+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -3,6 +3,7 @@ package roomescape.controller;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import roomescape.service.reservation.ReservationCreateService;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservation.dto.AdminReservationRequest;
 import roomescape.service.reservation.dto.ReservationFilterRequest;
@@ -15,15 +16,17 @@ import java.util.List;
 @RequestMapping("/admin/reservations")
 public class AdminReservationController {
     private final ReservationService reservationService;
+    private final ReservationCreateService reservationCreateService;
 
-    public AdminReservationController(ReservationService reservationService) {
+    public AdminReservationController(ReservationService reservationService, ReservationCreateService reservationCreateService) {
         this.reservationService = reservationService;
+        this.reservationCreateService = reservationCreateService;
     }
 
     @PostMapping
     public ResponseEntity<ReservationResponse> createReservation(
             @RequestBody @Valid AdminReservationRequest adminReservationRequest) {
-        ReservationResponse reservationResponse = reservationService.createAdminReservation(adminReservationRequest);
+        ReservationResponse reservationResponse = reservationCreateService.createAdminReservation(adminReservationRequest);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
                 .body(reservationResponse);
     }
