+++ src/main/java/roomescape/controller/ReservationController.java
@@ -5,6 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import roomescape.auth.LoginMemberId;
 import roomescape.service.auth.AuthService;
+import roomescape.service.reservation.ReservationCreateService;
 import roomescape.service.reservation.ReservationService;
 import roomescape.service.reservation.dto.ReservationRequest;
 import roomescape.service.reservation.dto.ReservationResponse;
@@ -16,10 +17,12 @@ import java.util.List;
 @RequestMapping("/reservations")
 public class ReservationController {
     private final ReservationService reservationService;
+    private final ReservationCreateService reservationCreateService;
     private final AuthService authService;
 
-    public ReservationController(ReservationService reservationService, AuthService authService) {
+    public ReservationController(ReservationService reservationService, ReservationCreateService reservationCreateService, AuthService authService) {
         this.reservationService = reservationService;
+        this.reservationCreateService = reservationCreateService;
         this.authService = authService;
     }
 
@@ -32,7 +35,7 @@ public class ReservationController {
     public ResponseEntity<ReservationResponse> createReservation(
             @RequestBody @Valid ReservationRequest reservationRequest,
             @LoginMemberId long memberId) {
-        ReservationResponse reservationResponse = reservationService.createMemberReservation(reservationRequest, memberId);
+        ReservationResponse reservationResponse = reservationCreateService.createMemberReservation(reservationRequest, memberId);
         return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
                 .body(reservationResponse);
     }
